requirements.txt
streamlit
PyMuPDF
pytesseract
pdf2image
pillow
pandas
